This will be quirky to get right.
Signal handling might be the way here?

+++ rsound/pcm_rsound.c
@@ -49,9 +49,7 @@ static snd_pcm_sframes_t rsound_write( snd_pcm_ioplug_t *io,
    buf = (char*)areas->addr + (areas->first + areas->step * offset) / 8;
 
    ssize_t result;
-   //fprintf(stderr, "doing fill_buffer: %d bytes\n", (int)size);
    result = fill_buffer(rsound, buf, size);
-   //fprintf(stderr, "ended fill_buffer: %d bytes\n", (int)size);
    if ( result <= 0 )
    {
       rsound_stop(io);
@@ -66,7 +64,8 @@ static snd_pcm_sframes_t rsound_pointer(snd_pcm_ioplug_t *io)
    int ptr;
    
    ptr = get_ptr(rsound);	
-   if ( ptr > rsound->alsa_buffer_size )
+
+   if ( ptr > (int)rsound->alsa_buffer_size )
       ptr = rsound->alsa_buffer_size;
    ptr = snd_pcm_bytes_to_frames( io->pcm, ptr );
    return ptr;
@@ -169,8 +168,6 @@ static int rsound_hw_params(snd_pcm_ioplug_t *io,
 	rsound->alsa_buffer_size *= rsound->bytes_per_frame;
    rsound->buffer_size = rsound->alsa_buffer_size;
 	rsound->alsa_fragsize *= rsound->bytes_per_frame;
-
-   //fprintf(stderr, "Buffer %d frag %d\n", (int)rsound->alsa_buffer_size, (int)rsound->alsa_fragsize);
 
    return 0;
 }
